% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_departure_hour.R
\name{allocate_departure_hour}
\alias{allocate_departure_hour}
\title{Sample starting hour for daily truck trips}
\usage{
allocate_departure_hour(local_daily_trips = NULL,
  regional_daily_trips = NULL, temporal_distributions, period_definitions,
  save_to = NULL)
}
\arguments{
\item{local_daily_trips}{Data frame containing local (intra-regional) daily
truck trips (optional, but only if regional_data_trips is defined)}

\item{regional_daily_trips}{Data frame containing regional (inter-regional)
daily truck trips (optional, but only if local_data_trips is defined)}

\item{temporal_distributions}{Data frame containing frequencies or
probabilities of trips by hour or period of the day, by truck type}

\item{period_definitions}{Data frame containing the period code corresponding
to each of the hours defined in temporal_distributions}

\item{save_to}{File name for saving the trip list with starting hour or
period appended}
}
\description{
Sample starting hour for daily truck trips
}
\details{
This function samples a user-specified distribution of frequencies
  or probabilities of trip starting hour or period for each trip in the input
  list. The user can pass either the local or regional trip list, or both,
  enabling the use of different temporal distributions for each group of
  trips if desired. Note that missing hours in the temporal distributions are
  not imputed; if there are no data for a given time slice no trips will be
  assigned to it. The resulting trip records are written with the starting
  hour or period appended to them.
}
\examples{
hourly_trips <- allocate_departure_hour(local_daily_trucks, NULL,
  temporal_distributions, period_definitions, save_to = "hourly_trucks.csv")
}
