% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocate_faf4_truckloads.R
\name{allocate_faf4_truckloads}
\alias{allocate_faf4_truckloads}
\title{Allocate flows from FAF regions to traffic analysis zones}
\usage{
allocate_faf4_truckloads(faf_truck_flows, internal_regions,
  zonal_attractions, zonal_equivalencies)
}
\arguments{
\item{faf_truck_flows}{Data frame containing truck flows between domestic
FAF regions}

\item{internal_regions}{List of FAF regions included within the modeled area}

\item{zonal_attractions}{Data frame containing number of attractions by truck
type by traffic analysis zone}

\item{zonal_equivalencies}{Data frame containing the mapping from FAF regions
to traffic analysis zones}
}
\description{
Allocate flows from FAF regions to traffic analysis zones
}
\details{
This function maps the domestic FAF regions coded in their regional
  flow database to traffic analysis zone (TAZ) equivalents. This can be done
  on a one-to-one basis in `zonal_equivalncies` when there is one and only
  one TAZ associated with a given FAF region. However, the FAF regions within
  the interior of the modeled area, defined in `internal_regions`, should be
  coded as missing values. They should be coded as `zonal_attractions`
  instead, where several TAZs can be defined for each internal FAF region.
  The resulting data frame includes all of the fields included in the input
  `faf_truck_flows` data frame with origin and destination fields added.
}
\examples{
exports <- allocate_faf4_truckloads(daily_trucks, 160, zonal_attractions,
  zonal_equivalencies)
exports <- allocate_faf4_truckloads(daily_trucks, c(411, 419, 532),
  zonal_attractions, zonal_equivalencies)
}
