% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_faf4_annual_truckloads.R
\name{create_faf4_annual_truckloads}
\alias{create_faf4_annual_truckloads}
\title{Create annual truckload equivalents from FAF v4+ regional flow database}
\usage{
create_faf4_annual_truckloads(faf_flows, load_consolidation_factor = 1)
}
\arguments{
\item{faf_flows}{Data frame containing the FAF interregional flows for the
analysis year, usually derived by applying preprecessor function(s) on
FHWA database for the desired target year.}

\item{load_consolidation_factor}{Numeric value representing the potential for
two or more commodity flows to be consolidated into single truck for
delivery (default = 1.0)}
}
\description{
Create annual truckload equivalents from FAF v4+ regional flow database
}
\details{
This function converts the commodity flows, measured in annual tons
  and value between FAF regions, into truckload equivalents. The five-step
  method described in the "FAF4 Freight Traffic Analysis" report is
  currently implemented, and the user should consult that documentation to
  understand the data requirements and methods used. The translation is
  deterministic, for it uses tables of proportions by vehicle type, body
  style, commodity, and trade type to arrive at one or more discrete annual
  truck trips by each of those indices, between each set of FAF regions that
  have commodity flows coded in the data. The result is a new, and much
  larger, data frame containing the number of truck trips for each inter-
  regional flow pair in the FAF data.
}
\examples{
annual_trucks <- create_faf4_annual_truckloads(faf_database)
}
