% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_faf_annual_truckloads.R
\name{create_faf_annual_truckloads}
\alias{create_faf_annual_truckloads}
\title{Create annual truckload equivalents from FAF v4+ regional flow database}
\usage{
create_faf_annual_truckloads(faf_database, truck_allocation_factors,
  truck_equivalency_factors, empty_truck_factors, target_year,
  method = "battelle", forecast_range_threshold = 5, save_to = NULL)
}
\arguments{
\item{faf_database}{Data frame containing the FAF regional full database in
special format used by this package, created by import_faf_database() or
comparable utility. The user can optionally provide the name (with path) of
comma-separated value (CSV) file containing the data frame.}

\item{truck_equivalency_factors}{Data frame containing truck equivalency
factors, or a string with the filename of a CSV file containing them.}

\item{empty_truck_factors}{Data frame containing empty truck factors, or a
string with the filename of a CSV file containing them.}

\item{target_year}{The four-digit simulation year as an integer value. The
value must be one of the years in the FAF database, or fall within the
interval x years before first year to x years after the time series in the
FAF data, where x is the foreast_range_threshold.}

\item{method}{The method used to map the tonnages to truckload equivalents
can be specified. The method devised by Battelle for use with FAF version
3.5 is the only one currently available.}

\item{forecast_range_threshold}{The number of years outside of the years
included in the FAF database that will be mapped to those those years. The
default is five years.}

\item{save_to}{File name for saving the annual truckload equivalents
generated by this program. These data are sometimes saved to reduce run
times for repeated runs of each simulation year, as well as in model
development and testing.}

\item{truck_allocaton_factors}{Data frame containing truck allocation
factors, of a string with the filename of a CSV file containing them.}
}
\description{
Create annual truckload equivalents from FAF v4+ regional flow database
}
\details{
This function converts the commodity flows, measured in annual tons
  and value between FAF regions, into truckload equivalents. The five-step
  method described in the "FAF3 Freight Traffic Analysis" report is
  currently implemented, and the user should consult that documentation to
  understand the data requirements and methods used. The translation is
  deterministic, for it uses tables of proportions by vehicle type, body
  style, commodity, and trade type to arrive at one or more discrete annual
  truck trips by each of those indices, between each set of FAF regions that
  have commodity flows coded in the data. The result is a new, and much
  larger, data frame containing the number of truck trips for each inter-
  regional flow pair in the FAF data.
}
\examples{
annual_trucks <- create_faf_annual_truckloads(faf_database,
 truck_allocation_factors, truck_equivalency_factors, empty_truck_factors,
 method = "Battelle", year = 2012)
}
