% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_contents.R
\name{data_frame_contents}
\alias{data_frame_contents}
\title{List the class and range of values in each column of a data frame}
\usage{
data_frame_contents(df, threshold = 10)
}
\arguments{
\item{df}{Data frame containing the variables to be summarized}

\item{threshold}{The number of discrete values of a certain variable to
enumerate. Defaults to 10, which typically fits within margin of knitr
output}
}
\description{
List the class and range of values in each column of a data frame
}
\details{
This function creates a data frame with the name, class, and range
  of values included for each column in a data frame, as well as the number
  of missing and zero values. It returns a data frame with that information,
  with one variable per record.
}
\examples{
df_contents <- data_frame_contents, 9)
}
