% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump_ISTDM_shapefile.R
\name{dump_ISTDM_shapefile}
\alias{dump_ISTDM_shapefile}
\title{Transform Cube assignment results in ESRI shapefile to data frame}
\usage{
dump_ISTDM_shapefile(shapefile, period_label, districts_def,
  areatypes = c("CBD", "Urban", "Rural"), combine.directions = TRUE,
  count.direction = "two-way")
}
\arguments{
\item{shapefile}{Data frame containing the contents of the dBASE (.dbf) file
created by Cube to export loaded network (traffic assignment) results.
Alternatively a string that contains a fully qualified filename (i.e.,
including full path information) can be supplied, which the function will
preprocess before proceeding.}

\item{period_label}{String denoting the period of the day that the data in
`shapefile` represents. This field will be added to the data frame.}

\item{districts_def}{Data frame containing a `DistrictID` for each
combination of `A` and `B` in `shapefile`.}

\item{areatypes}{A list of area type descriptions whose position in the list
correspond to integer values of `AT` in `shapefile`,}

\item{combine.directions}{Boolean variable denoting whether data for both
directions of two-day directional links will be combined to create a single
bidirectional link record (defaults to TRUE)}

\item{count.direction}{String denoting whether the daily counts coded in the
`shapefile` are one-way or two-way (bidirectional) counts (defauls to
`two-way`)}
}
\description{
Transform Cube assignment results in ESRI shapefile to data frame
}
\details{
This function converts loaded network data included in a dBASE
  (.dbf) file to those required for post-assignment summaries. The format is
  specific to the Idaho statewide travel model. A data frame is returned that
  includes variables that will be used in subsequent analyses.
}
\examples{
assignment_results <- dump_ISTDM_shapefile(shapefileFN, "AM",
  district_definitions, c("CBD", "Urban", "Rural"), TRUE, "two-way")
}
