% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_runtime_parameters.R
\name{get_runtime_parameters}
\alias{get_runtime_parameters}
\title{Create database of runtime parameters}
\usage{
get_runtime_parameters(filename, tokens_sep = "=", echo_parameters = TRUE)
}
\arguments{
\item{filename}{Fully qualified path and filename of text file containing
the user-defined runtime parameters.}

\item{tokens_sep}{Character value used in the parameters file to separate
key names and their respective values (defaults to equal sign)}

\item{echo_parameters}{Boolean flag specifying whether the contents of the
runtime parameters file should be printed to the console (default is TRUE)}
}
\description{
Create database of runtime parameters
}
\details{
This function reads a text file containing runtime parameters that
  will be globally available during the simulation. Each key-value pair must
  be specified on a single line. The key names are normally nor quoted, and
  string values generally do not need to be enclosed in quotes unless the
  string includes whitespace. The key-value pairs are stored in a global 
  list named RTP. It is up to the user to define each of the required run-
  time parameters, which depend upon the implemented model structure.
}
\examples{
RTP <<- get_runtime_parameters("~/Models/test/prelim_params.txt")
}
