% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_faf_database.R
\name{import_faf_database}
\alias{import_faf_database}
\title{Import FHWA Freight Analysis Framework (FHWA) Version 4.0+ regional database}
\usage{
import_faf_database(faf_data, distances = NULL, save_to = NULL)
}
\arguments{
\item{faf_data}{Data frame containing the FAF regional full database in the
format originally supplied by FHWA. Alternatively, the user can code the
filename (with path) of a comma-separated value (CSV) file containing the
same information, which the program will import into a data frame.}

\item{save_to}{File name for saving the FAF data transformed into the format
used internally by the CV modeling system}

\item{distance}{Data frame containing the inter-regional (FAF region to FAF
region) average distances. Alternatively, the user can code the filename
(with path) of a comma-separated value file containing the data.}
}
\description{
Import FHWA Freight Analysis Framework (FHWA) Version 4.0+ regional database
}
\details{
This function reads the FHWA Freight Analysis Framework Verion 4
  regional database in comma-separated value (CSV) format, as found on their
  website, and translates it into the format used in our analyses. Some
  categorical variables are mapped from integer values to more descriptive
  strings, value and tons are scaled back to real numbers, and average inter-
  regional distances are appended to each record. A single data frame of
  processed records is returned. It is important to note that the origin(s)
  and destination(s) are still coded in FAF regions at this point, as data
  still represent annual flows.
}
\examples{
faf43 <- import_faf_database(faf_flows, faf_distances)
faf43 <- import_faf_database("/Models/CT_Future2040/inputs/faf4_3_csv.zip",
  "faf_interregional_distances.csv.gz", "faf43-reformmatted.csv")
}
