% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Context-sensitive file reader}
\usage{
read_file(filename, format = "unknown")
}
\arguments{
\item{filename}{Fully qualified path and filename of data file to be read.
Both compresed and uncompressed data can be read, although it is assumed
that the underlying data are in comma-separated value (CSV) of feather
formats. The type of file is inferred from the file extension.}

\item{format}{String describing the format of the file, used in cases where
the file encoding cannot be inferred from the file extension(s). 
Defaults to "unknown".}
}
\description{
Context-sensitive file reader
}
\details{
This function is a general purpose file reader for data in comma-
  separated value (CSV) format. The goal is for the user to be able to
  specify the filename containing the data without having to worry about 
  which reader to use or handle different behaviors of specific readers.
  If the file extension is supported "it just works" without casting or
  wrapping the reader. In current version it handles feather or uncompressed
  CSV files, as well as those with gzip, zip, or xz compression. The data
  are returned in a data frame or tibble (depending upon the reader).
}
\examples{
trip_list <- read_file("combined_trips.csv.gz")
trip list <- read_file("https:://faf.fhwa.gov/faf44.zip", format = "csv")
}
