% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_idaho.R
\name{run_idaho}
\alias{run_idaho}
\title{Run pcvmodr for Idaho statewide model}
\usage{
run_idaho(runtime_parameters)
}
\arguments{
\item{runtime_parameters}{Fully-qualified path and filename of text file
containing runtime parameters required to run models built using pcvmodr
functions}
}
\description{
Run pcvmodr for Idaho statewide model
}
\details{
This function runs the interregional and local truck models built
  using pcvmodr functions for the Idaho statewide model.
}
\examples{
run_idaho("./runtime_parameters.txt")
}
