% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_faf4_daily_trucks.R
\name{sample_faf4_daily_trucks}
\alias{sample_faf4_daily_trucks}
\title{Sample daily FAF trucks from annual truckload estimates}
\usage{
sample_faf4_daily_trucks(annual_truckloads, target.week = 24,
  target.day = 4)
}
\arguments{
\item{annual_truckloads}{Data frame containing FAF flow records with 
estimates of annaul truckload equivalents appended.}

\item{target.week}{Integer value specifying the week of the year the
simulation should sample for (defaults to 24, typically occurs in
April)}

\item{target.day}{Integer value specifying the day of the week trips
should be sampled for (defaults to 4, which corresponds to 
Wednesday in the target.week)}
}
\description{
Sample daily FAF trucks from annual truckload estimates
}
\details{
This function creates discrete truck trip records from annual
  truckloads estimated eaelier, randomly assigns a travel week and day, and
  then saves those records that correspond to the user-specified simulation
  week and day. The user can specify observed flow distributions by week
  and day if desired, or sample from uniform distribution. The resulting
  trip records are for individual trucks, and include the FAF attributes
  associated with them. Note that the value and tonnage estimates have been
  split evenly across each of the trucks associatd with a given FAF flow
  record.
}
\examples{
daily_trucks <- sample_faf4_daily_trucks(faf_flows)
}
