% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_faf_daily_trucks.R
\name{sample_faf_daily_trucks}
\alias{sample_faf_daily_trucks}
\title{Sample discrete daily trucks from database of FAF annual truckload equivalents}
\usage{
sample_faf_daily_trucks(annual_truckload_equivalents, scaling_factor = 1,
  target_week = 19, target_day = 4, save_to = NULL)
}
\arguments{
\item{annual_truckload_equivalents}{Data frame containing the FAF regional
database with number of annual truck trips appended. If a string is passed
instead of data frame the program will attempt to read the data in expected
format from the full path and filename in the string.}

\item{scaling_factor}{is a floating point value that scales the annual trucks
to account for multiple stops and trip durations longer than the simulation
period (see description). The default value is 1.0.}

\item{target_week}{is an integer that defines which of the 52 weeks in the
year the trips are sampled from. The default value is 19, which roughly
corresponds in the second week in May.}

\item{target_day}{is an integer defining the day of the week trips are
included for. Note that this denotes the day the trip started, which is not
necessarily the day it ends or when it will be within the modeled area.
However, the latter is assumed to average out, obviating the need to
explicitly represent trip duration in the model. The default value is 4,
corresponding to Wednesday.}

\item{save_to}{File name for saving the daily interregional truck trips
created by this program.}
}
\description{
Sample discrete daily trucks from database of FAF annual truckload equivalents
}
\details{
This function samples weekly, and then daily, discrete truck trips
  based upon the number of annual truckload equivalents calculated for (FAF)
  or coded on (Transearch) each commodity flow record. The user can specify
  the week of the year and day of the week, although neither vary except
  randomly at the present. The user can also scale the annual trips to
  account for multiple stops per interregional trip and the fact that some
  (many?) FAF truck flows are longer than one day in duration. The resulting
  trip records have the FAF values appended, and the value and tons for each
  truck are split among all of the annual trucks.
}
\examples{
daily_trucks <- sample_faf_daily_trucks(annual_trucks)  # Accept defaults
daily_trucks <- sample_faf_daily_trucks(annual_truckloads,
 scaling_factor = 2.41, target_week = 24, save_to = "daily_trucks.csv")
}
