% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_local_destinations.R
\name{sample_local_destinations}
\alias{sample_local_destinations}
\title{Sample destinations for trips in local truck tours}
\usage{
sample_local_destinations(truck_origins, skim_distances,
  trip_length_targets, max_distance = 120, allow_intrazonals = TRUE,
  debug_tracing = FALSE)
}
\arguments{
\item{truck_origins}{Data frame containing discrete trip records with origin
zones annotated}

\item{skim_distances}{Data frame containing distances between each IDSTM_TAZ
combination}

\item{trip_length_targets}{Data frame containing observed or asserted trip
length frequencies or probabilities by distance for each truck type in the
simulation}

\item{max_distance}{The maximum distance that a local trip will travel to
reach a discrete destination (default = 120 miles)}

\item{allow_intrazonals}{A flag to indicate whether intrazonal trips will be
included in the data (default = TRUE)}

\item{debug_tracing}{A flag to indicate whether tracing statements for each
origin and truck type combination are printed (default = FALSE)}
}
\description{
Sample destinations for trips in local truck tours
}
\details{
This function carries out a singly-constrained destination choice,
  where the utility of each destination can be varied between distance and
  size term. The resulting tibble mirrors the input trip list with the
  addition of the destination.
}
\examples{
add_destinations <- sample_local_destinations(trip_records, skim_matrix,
  trip_length_targets)
}
