% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_temporal_allocation.R
\name{trip_temporal_allocation}
\alias{trip_temporal_allocation}
\title{Temporal allocation for individual trip records}
\usage{
trip_temporal_allocation(daily_trips, temporal_distributions)
}
\arguments{
\item{daily_trips}{Data frame containing the daily trip records}

\item{temporal_distributions}{Data frame containing the hourly shares for
trip starting times for each vehicle type defined in the simulation}
}
\description{
Temporal allocation for individual trip records
}
\details{
This function samples starting (departure) hour for each record in
  a trip list. Hourly shares must be specified for each vehicle type defined
  in the `temporal_distributions` data frame. The function does not ensure
  that all 24 hours in a day are defined, or even that values from 0-23 are
  required. However, at least one record for each vehicle type must be
  present. The resulting data frame adds `start_hour`, `period`, and `matrix`
  identifier to each record in the `daily_trips` data frame.
}
\examples{
add_start_hour <- trip_temporal_allocation(daily_truck_trips,
  observed_temporal_distributions)
}
