% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_zonal_attractions.R
\name{calculate_zonal_attractions}
\alias{calculate_zonal_attractions}
\title{Zonal attraction calculator}
\usage{
calculate_zonal_attractions(synthetic_firms, attraction_rates)
}
\arguments{
\item{synthetic_firms}{Data frame containing data for every synthetic firm
in the simulation.}

\item{attraction_rates}{Data frame containing truck trip generation rates for
each truck type used in the simulation, for each category of firms.}
}
\description{
Zonal attraction calculator
}
\details{
This function calculates zonal attractions by truck type for each
  synthetic firm using aggregate truck trip generation rates. The rates are
  defined by the user, and often come from the Quick Response Freight Manual
  (QRFM), ITE trip generation rates, or estimation results from a local or
  borrowed establishment survey. A data frame containing the attractions by
  these classification variables is returned
}
\examples{
attractions <- calculate_zonal_attractions(synthetic_firms, gen_rates)
}
